LPD-25944 Change components placement
