+++ modules/apps/frontend-data-set/frontend-data-set-views-web/types/src/main/resources/META-INF/resources/js/data_set/filters/modals/Configuration.d.ts
@@ -4,7 +4,7 @@
  */
 
 import React from 'react';
-import {IField, IFilter} from '../utils/types';
+import {IField, IFilter} from '../../../utils/types';
 interface IFilterModalConfigurationProps {
 	fieldInUseValidationError: boolean;
 	fieldNames?: string[];
