+++ modules/apps/frontend-data-set/frontend-data-set-views-web/types/src/main/resources/META-INF/resources/js/data_set/filters/modals/selection_filter/SelectionFilter.d.ts
